Crate sheave

Real Time Messaging Protocol

Mainly, this is the protocol for sending/receiving audio/video data over TCP.

Protocols

As the protocol based on RTMP, there are also following protocols:

RTMPE

RTMPS

RTMPT, RTMPTE, RTMPTS

RTMPE

Real Time Messaging Protocol encrypted. This will encrypt the chunk by using Diffie-Hellman key exchange. However this has been deprecated because this hadn't encrypted the network, that is, this can be targetted from man-in-the-middle attack.

RTMPS

Real Time Messaging Protocol over SSL/TLS. This will be secure than RTMPE because encrypts the connection with the client.

RTMPT, RTMPTE, RTMPTS

Real Time Messaging Protocol over HTTP/HTTPS. This is the same as RTMP, RTMPE and RTMPS except is on the HTTP/HTTPS.

Steps

RTMP takes following steps:

RTMP handshake

Invocations

Publishing

RTMP handshake

Takes the handshake with the client as RTMP. This can need HMAC-SHA256 digests/signatures at this moment. And if you will use RTMPS/RTMPTS, also the SSL certificate will need. See handshake.rs for more detail about the handshake.

Invocations

Exchanges application information to use, with the client. See handlers.rs for more detail about this phace, and also see messages.rs for more detail about each chunk.

Publishing

Stores audio/video data sent from the client. If some request from other clients has, sends their audio/video data to its client. See handlers.rs for more detail about this phase, and also see flv.rs for more detail about stored audio/video data.

Modules

decoders
The traits to encode the chunk messages.

encoders
The traits to encode the chunk messages.

errors
The errors

flv
The types for FLV file format.

handlers
The struct to handle RTMP

handshake
The structs/functions to do handshake for the RTMP layer.

messages
The chunk message patterns

Functions

run
Starts the server.