Trait sheave::decoders::GetByteBuffer

pub trait GetByteBuffer {
    fn get_u8(&mut self) -> Option<u8>;

    fn get_u16_be(&mut self) -> Option<u16>;

    fn get_u16_le(&mut self) -> Option<u16>;

    fn get_u24_be(&mut self) -> Option<u32>;

    fn get_u32_be(&mut self) -> Option<u32>;

    fn get_u32_le(&mut self) -> Option<u32>;

    fn get_f64(&mut self) -> Option<f64>;

    fn get_sliced_bytes(&mut self, len: usize) -> Option<Vec<u8>>;

    fn peek_byte(&self) -> Option<u8>;

    fn peek_bytes(&self, len: usize) -> Option<Vec<u8>>;
}

The byte getter.

The trait to get some byte from the buffer. This will require that considers the length to get or its byte order. The corresponding of the name to the length/the byte order is following:

Word	Length (in byte)	Byte order

_u8	1	Not considered

_u16_be	2	Big Endian

_u16_le	2	Little Endian

_u24_be	3	Big Endian

_u32_be	4	Big Endian

_u32_le	4	Little Endian

_f64	8	IEEE 754

Why have "get_" and "peek_" existed both?

The value what must check beforehand will exist. For exmaple, the object end marker of the AMF is 3 bytes, and is consisted of the value of 0x09 (just its marker) to follow after the empty 2 bytes. Note that we can regard their 2 bytes at the head as the string without its marker, but is unnecessary to insert them with the object end marker into the object. Because so, we must regard them as just empty 2 bytes, and must detect the object end marker as one including their 2 bytes. In this case, we will detect them as 0x009 without changing next reading position, then will terminate to decode for the object.

Therefore the method has been separated its name into "get_" and "peek_". This trait expects that you satisfy following things:

The method to start with "get_": decodes bytes then changes next reading position.

The method to start with "peek_": decodes bytes but don't change next reading position.

Required methods

fn get_u8(&mut self) -> Option<u8>

Takes just 1 byte from the buffer. If the 1 byte could be read, returns it after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u8(v: &Vec<u8>, offset: &mut usize) -> Option<u8> {
    v.get(*offset).map(
        |byte| {
            *offset +=  1;
            *byte
        }
    )
}

fn get_u16_be(&mut self) -> Option<u16>

Takes 2 bytes as Big Endian from the buffer. If the 2 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u16_be(v: &Vec<u8>, offset: &mut usize) -> Option<u16> {
    v.get(*offset..(*offset + 2)).map(
        |bytes| {
            *offset += 2;

            let mut array: [u8; 2] = [0; 2];

            array[0] = bytes[0];
            array[1] = bytes[1];
            u16::from_be_bytes(array)
        }
    )
}

fn get_u16_le(&mut self) -> Option<u16>

Takes 2 bytes as Little Endian from the buffer. If the 2 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u16_le(v: &Vec<u8>, offset: &mut usize) -> Option<u16> {
    v.get(*offset..(*offset + 2)).map(
        |bytes| {
            *offset += 2;

            let mut array: [u8; 2] = [0; 2];

            array[0] = bytes[0];
            array[1] = bytes[1];
            u16::from_le_bytes(array)
        }
    )
}

fn get_u24_be(&mut self) -> Option<u32>

Takes 3 bytes as Big Endian from the buffer. If the 3 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u24_be(v: &Vec<u8>, offset: &mut usize) -> Option<u32> {
    v.get(*offset..(*offset + 3)).map(
        |bytes| {
            *offset += 3;

            // Note that we must keep the first byte empty because are required to pass as u32.
            let mut array: [u8; 4] = [0; 4];

            array[1] = bytes[0];
            array[2] = bytes[1];
            array[3] = bytes[2];
            u32::from_be_bytes(array)
        }
    )
}

fn get_u32_be(&mut self) -> Option<u32>

Takes 4 bytes as Big Endian from the buffer. If the 4 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u32_be(v: &Vec<u8>, offset: &mut usize) -> Option<u32> {
    v.get(*offset..(*offset + 4)).map(
        |bytes| {
            *offset += 4;

            let mut array: [u8; 4] = [0; 4];

            array[0] = bytes[0];
            array[1] = bytes[1];
            array[2] = bytes[2];
            array[3] = bytes[3];
            u32::from_be_bytes(array)
        }
    )
}

fn get_u32_le(&mut self) -> Option<u32>

Takes 4 bytes as Little Endian from the buffer. If the 4 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_u32_le(v: &Vec<u8>, offset: &mut usize) -> Option<u32> {
    v.get(*offset..(*offset + 4)).map(
        |bytes| {
            *offset += 4;

            let mut array: [u8; 4] = [0; 4];

            array[0] = bytes[0];
            array[1] = bytes[1];
            array[2] = bytes[2];
            array[3] = bytes[3];
            u32::from_le_bytes(array)
        }
    )
}

fn get_f64(&mut self) -> Option<f64>

Takes 8 byte as IEEE 754 from the buffer. If the 8 bytes could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn get_f64(v: &Vec<u8>, offset: &mut usize) -> Option<f64> {
    v.get(*offset..(*offset + 8)).map(
        |bytes| {
            *offset += 8;

            let mut array: [u8; 8] = [0; 8];

            array[0] = bytes[0];
            array[1] = bytes[1];
            array[2] = bytes[2];
            array[3] = bytes[3];
            array[4] = bytes[4];
            array[5] = bytes[5];
            array[6] = bytes[6];
            array[7] = bytes[7];
            // Note that the f64 type can converted from only the u64 type.
            f64::from_bits(u64::from_be_bytes(array))
        }
    )
}

fn get_sliced_bytes(&mut self, len: usize) -> Option<Vec<u8>>

Takes the bytes just as much as what you specified. If they could be read, returns them after containing into the Option type, otherwis returns the None. This expects that you write the code equivalent to following:

fn get_sliced_bytes(v: &Vec<u8>, offset: &mut usize, len: usize) -> Option<Vec<u8>> {
    v.get(*offset..(*offset + len)).map(
        |bytes| {
            *offset += len;
            bytes.to_vec()
        }
    )
}

Parameters

len: usize

The length to take bytes from the buffer.

fn peek_byte(&self) -> Option<u8>

Takes 1 byte without changing the current position. If the 1 byte could be read, returns it after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn peek_byte(v: &Vec<u8>, offset: usize) -> Option<u8> {
    // Note that you shouldn't change the current position in this case.
    v.get(offset).map(
        |byte| *byte
    )
}

fn peek_bytes(&self, len: usize) -> Option<Vec<u8>>

Takes the bytes just as much as what you specified without changing the current position. If they could be read, returns them after containing into the Option type, otherwise returns the None. This expects that you write the code equivalent to following:

fn peek_bytes(v: &Vec<u8>, offset: usize, len: usize) -> Option<Vec<u8>> {
    // Note that you shouldn't change the current position in this case.
    v.get(offset..(offset + len)).map(
        |bytes| {
            bytes.to_vec()
        }
    )
}

Parameters

len: usize

The length to take bytes from the buffer.

Loading content...