Trait sheave::decoders::RtmpDecoder

pub trait RtmpDecoder: GetByteBuffer {
    fn decode_basic_header(&mut self) -> Option<BasicHeader>;

    fn decode_message_header(
        &mut self, 
        message_format: MessageFormat
    ) -> Option<MessageHeader>;

    fn decode_extended_timestamp(
        &mut self, 
        message_header: MessageHeader
    ) -> Option<Duration>;

    fn decode_chunk_size(&mut self) -> Option<ChunkData>;

    fn decode_bytes_read(&mut self) -> Option<ChunkData>;

    fn decode_ping(&mut self) -> Option<ChunkData>;

    fn decode_server_bandwidth(&mut self) -> Option<ChunkData>;

    fn decode_client_bandwidth(&mut self) -> Option<ChunkData>;

    fn decode_audio(&mut self) -> Option<ChunkData>;

    fn decode_video(&mut self) -> Option<ChunkData>;

    fn decode_amf_number(&mut self) -> Option<AmfData>;

    fn decode_amf_boolean(&mut self) -> Option<AmfData>;

    fn decode_amf_string(&mut self) -> Option<AmfData>;

    fn decode_amf_object(&mut self) -> Option<AmfData>;

    fn decode_amf_null(&mut self) -> Option<AmfData>;

    fn decode_amf_mixed_array(&mut self) -> Option<AmfData>;

    fn decode_amf_data(&mut self) -> Option<AmfData>;

    fn decode_notify(&mut self) -> Option<ChunkData>;

    fn decode_invoke_connect_result(&mut self) -> Option<ChunkData>;

    fn decode_invoke_release_stream_result(&mut self) -> Option<ChunkData>;

    fn decode_invoke_create_stream_result(&mut self) -> Option<ChunkData>;

    fn decode_invoke_on_fc_publish(&mut self) -> Option<ChunkData>;

    fn decode_invoke_on_status(&mut self) -> Option<ChunkData>;

    fn decode_invoke(&mut self) -> Option<ChunkData>;

    fn decode_unknown(&mut self) -> Option<ChunkData>;

    fn decode_chunk_data(
        &mut self, 
        message_type: MessageType
    ) -> Option<ChunkData>;
}

The decoder for the RTMP chunk

The trait to decode the bytes into the RTMP chunk.

Required methods

fn decode_basic_header(&mut self) -> Option<BasicHeader>

Decodes the chunk basic header. This expects that you satisfy following points:

You regard first 2 bits as the format of the chunk message header.

If next 6 bits are 0, you regard next 1 byte as the chunk stream id.

If they are 1, you regard next 2 bytes as the chunk stream id.

Note that you must decode them as Little Endian.

Otherwise you regard its 6 bits as the chunk stream id.

If you couldn't decode above fields all, return the None.

fn decode_message_header(
&mut self,
message_format: MessageFormat
) -> Option<MessageHeader>

Decodes the chunk message header. This expects that you satisfy following points:

You regard first 3 bytes as the timestamp.

You regard next 3 bytes as the message length.

You regard next 1 byte as the message type id.

You regard next 4 bytes as the message stream id.

Note that you must decode them as Little Endian.

If you couldn't decode above fields all, return the None.

Parameters

message_format: MessageFormat

The chunk message header's format. You can get this by using decode_basic_header() to the buffer. The returned MessageHeader will rely on the value got from above method for its format.

fn decode_extended_timestamp(
&mut self,
message_header: MessageHeader
) -> Option<Duration>

Decodes the extended timestamp. This expects that you satisfy following points:

If the timestamp in the chunk message header hasn't exceeded 0xffffff, returns the None.

Otherwise you regard the 4 bytes as the extended timestamp.

You check that the timestamp in the chunk message header has been 0xffffff.

Though it will be improbable, you set 0xffffff to the timestamp in the chunk message header forcibly if it hasn't been 0xffffff.

If you couldn't decode the extended timestamp, return the None.

Parameters

message_header: MessageHeader

The message header decoded just before. It will rely on the timestamp in this value for whether the extended timestamp will exist.

fn decode_chunk_size(&mut self) -> Option<ChunkData>

Decodes the chunk data as Chunk Size. This expects that you satisfy following points:

You regard the 4 bytes as Chunk Size's value.

If the most significant bit is 1, you emit the panic!.

If you couldn't decode Chunk Size, return the None.

fn decode_bytes_read(&mut self) -> Option<ChunkData>

Decodes the chunk data as Bytes Read. This expects that you satisfy following points:

You regard the 4 bytes as Bytes Read's value.

If you couldn't decode Bytes Read, return the None.

fn decode_ping(&mut self) -> Option<ChunkData>

Decodes the chunk data as Ping. This expects that you satisfy following points:

You regard first 2 bytes as Ping's event type.
Ping's event data will be associated with their event types.
See messages.rs more details about the correspondence of Ping's event type to its event data.

You regard bytes following after Ping's event type as its event data, and decode their event data corresponding to its event type.

If you couldn't decode either event type or event data, return the None.

fn decode_server_bandwidth(&mut self) -> Option<ChunkData>

Decodes the chunk data as Server Bandwidth. This expects that you satisfy following points:

You regard the 4 bytes as Server Bandwidth's value.

If you couldn't decode Server Bandwidth, return the None.

fn decode_client_bandwidth(&mut self) -> Option<ChunkData>

Decodes the chunk data as Client Bandwidth. This expects that you satisfy following points:

You regard first 4 bytes as Client Bandwidth's value.

You regard next 1 byte as the limit type of Client Bandwidth.
See messages.rs more details about its limit type.

If you couldn't decode either client bandwidth or limit type, return the None.

fn decode_audio(&mut self) -> Option<ChunkData>

Decodes the chunk data as Audio. This expects that you satisfy following points:

You regard remaining bytes as Audio data.

If you couldn't decode Audio data, return the None.

Note that you shouldn't consider about the file format and the codec in this moment. Because the program will become intricately.

fn decode_video(&mut self) -> Option<ChunkData>

Decodes the chunk data as Video. This expects that you satisfy following points:

You regard remaining bytes as Video data.

If you couldn't decode Video data, return the None.

Note that you shouldn't consider about the file format and the codec in this moment due to the same cause as Audio data.

fn decode_amf_number(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's number. This expect that you satisfy following points:

You regard the 8 bytes as a floating point number (IEEE 754).

If you couldn't decode them, return the None.

fn decode_amf_boolean(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's boolean. This expects that you satisfy following points:

You regard the 1 byte as a (AMF's) boolean.

If it is 0, you regard it as the false.

If it is else, you regard it as the true.

If you couldn't decode it, return the None.

fn decode_amf_string(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's string. This expects that you satisfy following points:

You regard first 2 bytes as its string's length.

You take actual string just as much as its length, then decode it as UTF-8.

If you took the string what its length is 0, may decode it into the Some("".to_string()).

If you couldn't decode above all, return the None.

fn decode_amf_object(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's object. This expects that you satisfy following points:

You must associate its name with its value.

The name will exist as the string without AMF's marker.

The value will exist to follow after its name.
This will have AMF's marker.

If you detected the object end marker following after the empty string key, you must terminate to decode AMF's object.
Or you may detect it as just 0x000009.

If you couldn't decode both the name and the value completely, return the None.

fn decode_amf_null(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's null. This expects that you satisfy following points:

You regard this as nothing to decode, and return the AmfData::Null.

fn decode_amf_mixed_array(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF's mixed array. This expects that you satisfy following points:

You regard first 4 bytes as the length of this array.

You regard remaining bytes as what the name and the value will follow after.

The length will become equal to the count of name/value pairs, though you may discard it because this array will also contain the object end marker.

If you couldn't decode both the length and the name/value pairs completely, return the None.

fn decode_amf_data(&mut self) -> Option<AmfData>

Decodes a part of the chunk data as AMF. This expects that you satisfy following points:

You regard first 1 byte as AMF's data type marker, and regard bytes following it as actual AMF's data.
See AmfData for more details about the correspondence of AMF's data type to actual value.

If you couldn't decode both the AMF data type and its data completely, return the None.

And keep your mind that you should use the method named as decode_amf_* when will decode bytes as AMF. The correspondence of the AMF data type to the method what should use when decode AMF's data bytes is following:

AMF data type (id)	Method

0	decode_amf_number()

1	decode_amf_boolean()

2	decode_amf_string()

3	decode_amf_object()

5	decode_amf_null()

8	decode_amf_mixed_array()

You will this method when will decode actual AMF data after checking id of the AMF data type.

fn decode_notify(&mut self) -> Option<ChunkData>

Decodes the chunk data as Notify. This expects that you satisfy following points:

You regard the chunk data as the AMF's MixedArray what will follow after 2 AMF's Strings.

First string will be "@setDataFrame".

Second string will be "onMetaData".

If you couldn't decode their strings and MixedArray, return the None.

fn decode_invoke_connect_result(&mut self) -> Option<ChunkData>

Decode the chunk data as the response message of the connect command of the NetConnection command. This expects that you satisfy following points:

You decode regarding the response message as what will have following 4 fileds:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the request of connect has (Number)

Properties: an AMF's Object

Information: See the InfoObject (Object)

If you couldn't decode above fields all, return the None.

fn decode_invoke_release_stream_result(&mut self) -> Option<ChunkData>

Decodes the chunk data as the response message of the releaseStream command of the NetConnection command. This expects that you satisfy following points:

You decode regarding the response message as what will have following 4 fields:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the releaseStream request has (Number)

Command object: probable nothing (Null)

If you couldn't decode above fields all, return the None.

fn decode_invoke_create_stream_result(&mut self) -> Option<ChunkData>

Decodes the chunk data as the response message of the createStream command of the NetConnection command. This expects that you satisfy following points:

You decode regarding the response message as what will have following 4 fields:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the createStream request has (Number)

Command object: probable nothing (Null)

Message stream id: one what the server is holding for this client (Number)

If you couldn't decode above field all, return the None.

fn decode_invoke_on_fc_publish(&mut self) -> Option<ChunkData>

Decodes the response message of the FCPublish command. This expects that you satisfy following points:

You decode regarding the response message as what will have following 1 field:

Command name: "onFCPublish" (String)

If you couldn't above field, return the None.

fn decode_invoke_on_status(&mut self) -> Option<ChunkData>

Decodes the response message of the NetStream command. This expects that you satisfy following points:

You deocode regarding the response message as what will have following 4 fields:

Command name: "onStatus" (String)

Transaction id: 0 (Number)

Command object: probable nothing (Null)

Info object: See the InfoObject (Object)

If you couldn't decode above field all, return the None.

fn decode_invoke(&mut self) -> Option<ChunkData>

Decodes the invocation request message. This expects that you satisfy following points:

You decode regarding the invocation request as what will have following messages:

"connect"

"releaseStream"

"FCPublish"

"createStream"

"publish"

And you decode regarding above messages as what will have following fields respectively:

connect

Command name: "connect" (String)

Transaction id: 1 (Number)

Command object: See the CommandObject (Object)

Optional argument: In the official specification paper, this has been specified the Object, though we can't check the existence of this field.
Therefore you won't be required to consider about this field.

releaseStream

Command name: "releaseStream" (String)

Transaction id: probable 2 (Number)

Command object: probable nothing (Null)

Playpath: playpath in "rtmp://example.com/appName/playpath" (if it is specified) (String)

FCPublish

Command name: "FCPublish" (String)

Transaction id: probable 3 (Number)

Command object: probable nothing (Null)

Playpath: playpath in "rtmp://example.com/appName/playpath" (if it is specified) (String)

createStream

Command name: "createStream" (String)

Transaction id: probable 4 (Number)

Command object: probable nothing (Null)

publish

Command name: "publish" (String)

Transaction id: probable 5 (Number)

Command object: probable nothing (Null)

Playpath: playpath in "rtmp://example.com/appName/playpath" (if it is specified) (String)

Playtype: either "live", "record" or "append" (String)

If you couldn't decode above messages, return the None

fn decode_unknown(&mut self) -> Option<ChunkData>

Decodes the chunk data as Unknown bytes. This expects that you satisfy following points:

You regard remaining bytes as some unknown bytes.
You aren't required to consider that their bytes is what in this moment.

You may pass them through other processes keeping them as be.

fn decode_chunk_data(&mut self, message_type: MessageType) -> Option<ChunkData>

Decodes the chunk data. This expects that you satisfy following points:

You must associate the message type id with actual chunk data.
See the MessageType and the ChunkData for more details about their correspondence.

If you couldn't decode the chunk data, return the None.

And keep your mind that you should use the method named as "decode_*" when will decode the chunk data. The correspondence of the message type and the method what should use when decode the chunk data is following:

Message type id	Method

1	decode_chunk_size()

3	decode_bytes_read()

4	decode_ping()

5	decode_server_bandwidth()

6	decode_client_bandwidth()

8	decode_audio()

9	decode_video()

18	decode_notify()

20	decode_invoke()

other	decode_unknown()

You will use this method when will decode actual chunk data after checking the message type id in the chunk message header.

Parameters

message_type: MessageType

The message type input in the chunk message header. The chunk data must be associated actual content with the message type ida in the chunk message header.

Loading content...