Trait sheave::encoders::PutByteBuffer

pub trait PutByteBuffer {
    fn put_u8(&mut self, byte: u8);

    fn put_u16_be(&mut self, byte: u16);

    fn put_u16_le(&mut self, byte: u16);

    fn put_u24_be(&mut self, byte: u32);

    fn put_u32_be(&mut self, byte: u32);

    fn put_u32_le(&mut self, byte: u32);

    fn put_f64(&mut self, byte: f64);

    fn put_bytes(&mut self, bytes: Vec<u8>);
}

The byte putter

The trait to put some value into the buffer. This will require that consider the length to put or its byte order. The correspondence of the name to the length/the byte order is following:

Word	Length (in byte)	Byte order

_u8	1	Not considered

_u16_be	2	Big Endian

_u16_le	2	Little Endian

_u24_be	3	Big Endian

_u32_be	4	Big Endian

_u32_le	4	Little Endian

_f64	8	IEEE 754

Required methods

fn put_u8(&mut self, byte: u8)

Puts just 1 byte into the buffer. This expects that you write the code equivalent to following:

fn put_u8(v: &mut Vec<u8>, byte: u8) {
    v.push(byte);
}

Parameters

byte: u8

A 1 byte value to put into the buffer.

fn put_u16_be(&mut self, byte: u16)

Puts 2 bytes as Big Endian into the buffer. This expects that you write the code equivalent to following:

fn put_u16_be(v: &mut Vec<u8>, byte: u16) {
    v.extend_from_slice(&byte.to_be_bytes());
}

Parameters

byte: u16

A 2 bytes value to put into the buffer. This is expected to convert into the [u8; 2] type as Big Endian.

fn put_u16_le(&mut self, byte: u16)

Puts 2 bytes as Little Endian into the buffer. This expects that you write the code equivalent to following:

fn put_u16_le(v: &mut Vec<u8>, byte: u16) {
    v.extend_from_slice(&byte.to_le_bytes());
}

Parameters

byte: u16

A 2 bytes value to put into the buffer. This is expected to convert into the [u8; 2] type as Little Endian.

fn put_u24_be(&mut self, byte: u32)

Puts 3 bytes as Big Endian into the buffer. This expects that you write the code equivalent to following:

fn put_u24_be(v: &mut Vec<u8>, byte: u32) {
    // The first byte won't be needed because this method is requiring just 3 bytes.
    v.extend_from_slice(&byte.to_be_bytes()[1..]);
}

Parameters

byte: u32

A 4 bytes value to put into the buffer. This is expected to convert into the [u8; 4] type as Big Endian. Though this is 4 bytes because the type for 3 bytes isn't existing yet, keep this value up to 0x00ffffff.

fn put_u32_be(&mut self, byte: u32)

Puts 4 bytes as Big Endian into the buffer. This expects that you write the code equivalent to following:

fn put_u32_be(v: &mut Vec<u8>, byte: u32) {
    v.extend_from_slice(&byte.to_be_bytes());
}

Parameters

byte: u32

A 4 bytes value to put into the buffer. This is expected to convert into the [u8; 4] type as Big Endian.

fn put_u32_le(&mut self, byte: u32)

Puts 4 bytes as Little Endian into the buffer. This expects that you write the code equivalent to following:

fn put_u32_le(v: &mut Vec<u8>, byte: u32) {
    v.extend_from_slice(&byte.to_le_bytes());
}

Parameters

byte: u32

A 4 bytes value to put into the buffer. This is expected to convert into the [u8; 4] type as Little Endian.

fn put_f64(&mut self, byte: f64)

Puts 8 bytes as IEEE 754 into the buffer. This expects that you write the code equivalent to following:

fn put_f64(v: &mut Vec<u8>, byte: f64) {
    // f64::to_bits() makes the u64 value from the f64 value.
    v.extend_from_slice(&byte.to_bits().to_be_bytes());
}

Parameters

byte: f64

A 8 bytes value to put into the buffer. This is expected to convert into the [u8; 8] type as Big Endian.

fn put_bytes(&mut self, bytes: Vec<u8>)

Puts the bytes into the buffer as it be. This expects that you write the code equivalent to following:

// Note that "mut" is a pattern, and can't use in the trait defenition.
fn put_bytes(v: &mut Vec<u8>, mut bytes: Vec<u8>) {
    v.append(&mut bytes);
}

Parameters

bytes: Vec<u8>

The bytes to put into the buffer.

Loading content...