Trait sheave::encoders::RtmpEncoder

pub trait RtmpEncoder: PutByteBuffer {
    fn encode_basic_header(&mut self, basic_header: BasicHeader);

    fn encode_message_header(&mut self, message_header: MessageHeader);

    fn encode_extended_timestamp(&mut self, timestamp: Option<Duration>);

    fn encode_chunk_size(&mut self, chunk_size: u32);

    fn encode_bytes_read(&mut self, bytes_read: u32);

    fn encode_ping(&mut self, ping_data: PingData);

    fn encode_server_bandwidth(&mut self, bandwidth: u32);

    fn encode_client_bandwidth(&mut self, bandwidth: u32, limit_type: LimitType);

    fn encode_audio(&mut self, bytes: Vec<u8>);

    fn encode_video(&mut self, bytes: Vec<u8>);

    fn encode_amf_number(&mut self, number: f64);

    fn encode_amf_boolean(&mut self, boolean: bool);

    fn encode_amf_string(&mut self, string: String);

    fn encode_amf_object(&mut self, object: HashMap<String, AmfData>);

    fn encode_amf_null(&mut self);

    fn encode_amf_mixed_array(&mut self, mixed_array: HashMap<String, AmfData>);

    fn encode_amf_data(&mut self, data: AmfData);

    fn encode_notify(&mut self, notify_command: NotifyCommand);

    fn encode_invoke_net_connection(
        &mut self, 
        net_connection: NetConnectionCommand
    );

    fn encode_invoke_net_stream(&mut self, net_stream: NetStreamCommand);

    fn encode_invoke_fc_publish(&mut self, fc_publish: FcPublishCommand);

    fn encode_invoke(&mut self, invoke: InvokeCommand);

    fn encode_unknown(&mut self, unknown: Vec<u8>);

    fn encode_chunk_data(&mut self, chunk_data: Option<ChunkData>);
}

The encoder for the RTMP chunk

The trait to encode the RTMP chunk into the bytes.

Required methods

fn encode_basic_header(&mut self, basic_header: BasicHeader)

Encodes the chunk basic header into the bytes. This expects that you satisfy following points:

You put this header into the buffer at head.

See messages.rs for more detail about the chunk basic header.

Parameters

basic_header: BasicHeader

The chunk basic header to put into the buffer. See BasicHeader for more detail about this type.

fn encode_message_header(&mut self, message_header: MessageHeader)

Encodes the chunk message header into the bytes. This expects that you satisfy following points:

You put this header into just after the chunk basic header in the buffer.

If this header is Format 3, you must input nothing. See messages.rs for more detail about the chunk basic header.

Parameters

message_header: MessageHeader

The chunk message header to put into the buffer. See MessageHeader for more detail about this type.

fn encode_extended_timestamp(&mut self, timestamp: Option<Duration>)

Encodes the extended timestamp into the bytes. This expects that you satisfy following points:

You this into just after the chunk message header in the buffer.

If this is None, you encode nothing.

Otherwise you put this into the buffer as 4 bytes value.

See messages.rs for more detail about the extended timestamp.

Parameters

timestamp: Option<Duration>

The extended timestamp to put into the buffer. Consider how to process the value above 0xffffffff.

fn encode_chunk_size(&mut self, chunk_size: u32)

Encodes the chunk size value into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

See messages.rs for more detail about the chunk size.

Parameters

chunk_size: u32

The chunk size to put into the buffer. Consider how to process the value above 0x7fffffff.

fn encode_bytes_read(&mut self, bytes_read: u32)

Decodes the size of total read bytes into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

See messages.rs for more detail about the size of read bytes.

Parameter

bytes_read: u32

The size of read bytes to put into the buffer.

fn encode_ping(&mut self, ping_data: PingData)

Encodes the ping event into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

You put Ping's data following after its event type (2 bytes).

See messages.rs for more detail about Ping.

Parameters

ping_data: PingData

Ping's data to put into the buffer. See PingType and PingData for more detail about this type.

fn encode_server_bandwidth(&mut self, bandwidth: u32)

Encodes server side bandwidth into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

See messages.rs for more detail about the server bandwidth.

Parameters

bandwidth: u32

The server side bandwidth to put into the buffer.

fn encode_client_bandwidth(&mut self, bandwidth: u32, limit_type: LimitType)

Encodes the client side bandwidth and its limit type into the bytes. This expects that you satisfy following points:

You put these into following places:

If the extended timestamp is existing, you put these into just after the extended timestamp in the buffer.

Otherwise you put these into just after the chunk message header in the buffer.

You put the limit type (1 byte) after the client side bandwidth.

See messages.rs for more detail about the client side bandwidth.

Parameters

bandwidth: u32

The client side bandwidh to put into the buffer.

limit_type: LimitType

The limit type for the client side bandwidth to put into the buffer. See LimitType for more detail about this type.

fn encode_audio(&mut self, bytes: Vec<u8>)

Encodes the audio data into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

Note that you shouldn't consider about the file format and the codec in this moment. Because the program will become intricately. See flv.rs for more detail about the audio data.

Parameter

bytes: Vec<u8>

The audio data to put into this buffer.

[flv.rs] ../flv.rs.html

fn encode_video(&mut self, bytes: Vec<u8>)

Encodes the video data into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

Note that you shouldn't consider about the file format and the codec in this moment. Because the program will become intricately. See flv.rs for more detail about the video data.

Parameters

bytes: Vec<u8>

The video data to put into the buffer.

fn encode_amf_number(&mut self, number: f64)

Encodes f64 value into the bytes of the AMF's Number. This expects that you satisfy following points:

You put this into the buffer following after an AMF data type marker of the number(0x00).

Parameters

number: f64

The number to put into the buffer.

fn encode_amf_boolean(&mut self, boolean: bool)

Encodes bool value into the bytes of AMF's Boolean. This expects that you satisfy following points:

You put this into the buffer following after an AMF data type marker of the boolean (0x01).

You regard false as 0, true as otherwise.

Parameters

boolean: bool

The boolean value to put into the buffer.

fn encode_amf_string(&mut self, string: String)

Encodes string value into the bytes of the AMF's String. This expects that you satisfy following points:

You put this into the buffer following after an AMF data type marker of the string (0x02).

At this moment, you must put the length of this string (2 bytes integer, Big Endian) before actual string.
The length must be equal to actual string length.

You must encode this sure as UTF-8.

Parameters

string: String

The string to put into this buffer. You make sure that this is encoded as UTF-8.

fn encode_amf_object(&mut self, object: HashMap<String, AmfData>)

Encodes hash map value into the bytes of AMF's Object. This expects that you satisfy following pints:

You put this into the buffer following after the AMF data type marker of the object (0x03).

At this moment, you make sure that the object has consisted of name/value pairs:

The name is a string without only the AMF data type marker.

The value is some AMF's value.

AMF's object must terminate by AMF's object end marker (0x09), and the object end marker must follow after a key of empty string.
That is, the object end marker will consist of 0x000009 (3 bytes).

You must encode the name as as UTF-8.

Parameters

object: HashMap<String, AmfData>

The object to put into the buffer. You make sure that the key is encoded as UTF-8.

fn encode_amf_null(&mut self)

Encodes the AMF's Null into the bytes. This expects that you satisfy following points:

You regard this as what has no value.
That is, you put only the AMF data type marker into the buffer.

fn encode_amf_mixed_array(&mut self, mixed_array: HashMap<String, AmfData>)

Encodes hash map value into the bytes of AMF's MixedArray. This expects that you satisfy following points:

You put this into the buffer following after the AMF data type marker of mixed array (0x08).

At this moment, you make sure that the mixed array has consisted of name/value pairs with their length.

The length is a 4 bytes integer (Big Endian).
The length must be equal to actual count of name/value pairs, though we can ignore this value because the mixed array has the object end marker.

The name is a string without only the AMF data type marker.

The value is some AMF's value.

AMF's mixed array must terminate by AMF's object end marker (0x09), and the object end marker must follow after a key of empty string.
That is, the object end marker will consist of 0x000009 (3 bytes).

You must encode the name as UTF-8.

Parameters

mixed_array: HashMap<String, AmfData>

The mixed array to put into the buffer. You make sure that the key is encoded as UTF-8.

fn encode_amf_data(&mut self, data: AmfData)

Encodes the AMF's value into the bytes. This expects that you satisfy following points:

You put this into the buffer following after its data type marker.
At this moment, you should use the method named as "encode_amf_*" corresponded to its data type marker.
The correspondence of the AMF data type to the method what should use when encode AMF's data bytes is following:

AMF data type	Method

Number	encode_amf_number()

Boolean	encode_amf_boolean()

String	encode_amf_string()

Object	encode_amf_object()

Null	encode_amf_null()

MixedArray	encode_amf_mixed_array()

See AmfDataType for more detail about the AMF data type marker.

Parameters

data: AmfData

Some AMF's value to put into the buffer. See AmfData for more detail about the AMF's value.

fn encode_notify(&mut self, notify_command: NotifyCommand)

Encodes the notify message into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

At this morment, you put following fields into the buffer respectively:

First, "@setDataFrame" (String).

Second, "onMetaData" (String).

Then, actual MetaData for FLV (MixedArray).

See AmfDataType and MetaData for more detail about the AMF data type marker and the content of actual metadata.

Parameters

notify_command: NotifyCommand

The content of notify message to put into this buffer. See AmfData and NotifyCommand for more detail about this type.

fn encode_invoke_net_connection(&mut self, net_connection: NetConnectionCommand)

Encodes the NetConnection command of the invocation message into the bytes. This expects that you satisfy following points:

You encode regarding the NetConnection command as what will have following messages:

"connect"

"releaseStream"

"createStream"

And you encode regarding above messages as what will have following fields respectively:

connect

In request:

Command name: "connect" (String).

Transaction id: 1 (Number).

Command object: See the CommandObject (Object)

Optional Argument: In the official specification paper, this has been specified the Object, though we can't check the existence of this field.
Therefore you won't be required to consider about this field.

In response:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the request of the connect has (Number)

Properties: an AMF's Object

Information: See the InfoObject (Object).

releaseStream

In request:

Command name: "releaseStream" (String)

Transaction id: probable 2 (Number)

Command object: See the CommandObject (Object).
Note that you must put an AMF's Null into the buffer instead if this is nothing.

Playpath: playpath in "rtmp://example.com/appName/playpath" (if is specified) (String)

In response:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the request of releaseStream has (Number)

Command object: See the CommandObject (Object).
Note that you must put AMF's Null into the buffer instead if this is nothing.

createStream

In request:

Command name: "createStream" (String)

Transaction id: probable 4 (Number)

Command object: see the CommandObject (Object).
Note that you must put AMF's Null into the buffer instead if this is nothing.

In response:

Command name: "_result" or "_error" (String)

Transaction id: the same value as what the request of createStream has (Number)

Command object: See the CommandObject (Object).
Note that you must put AMF's Null into the buffer instead if this is nothing.

Message stream id: new stream message id for its client (Number)

Parameters

net_connection: NetConnectionCommand

The NetConnection command of invocation message to put into the buffer. See the NetConnectionCommand for more detail about this type.

fn encode_invoke_net_stream(&mut self, net_stream: NetStreamCommand)

Encodes the NetStream command of the invocation message into the bytes. This expects that you satisfy following points:

You encode regarding the NetStream command as what will have following messages:

"publish"

And you encode regarding above messages as what will have following fields respectively:

publish

In request:

Command name: "publish" (String)

Transaction id: probable 5 (Number)

Command object: an AMF's Null

Playpath: playpath in "rtmp://example.com/appName/playpath" (if is specified) (String)

Playtype: either "live", "record" or "append" (String)

In response:

Command name: "onStatus" (String)

Transaction id: 0 (Number)

Command object: an AMF's Null

Info object: See the InfoObject (Object)

Parameters

net_stream: NetStreamCommand

The NetStream command to put into the buffer. See the NetStreamCommand for more detail about this type.

fn encode_invoke_fc_publish(&mut self, fc_publish: FcPublishCommand)

Encodes the FCPublish command of the invocation message into the bytes. This expects that you satisfy following points:

You encode regarding the FCPublish command as what will have following messages:

"FCPublish"

And you encode regarding above messages as what will have following fields respectively:

FCPublish

In request:

Command name: "FCPublish" (String)

Transaction id: probable 3 (Number)

Command object: an AMF's Null

Playpath: playpath in "rtmp://example.com/appName/playpath" (if is specified) (String)

In response:

Command name: "onFCPublish" (String)

Parameters

fc_publish: FcPublishCommand

The FCPublish command to put into the buffer. See the FcPublishCommand for more detail about this type.

fn encode_invoke(&mut self, invoke: InvokeCommand)

Encodes some invocation message into the bytes. This expects that you satisfy following points:

You put this into following place:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

At this moment, you put following fields into the buffer respectively:

Some command name of subcommands (String)

Transaction id (Number)

Actual data for subcommands (some AMF's value)

You should use the method named as "encode_invoke_*" corresponded to the type of the command.
The correspondence of the the command to the method what should use when encode the command is following:

Command	Method

NetConnection	encode_invoke_net_connect()

NetStream	encode_invoke_net_stream()

FCPublish	encode_invoke_fc_publish()

Parameters

invoke: InvokeCommand

The command of invocation message to put into the buffer. See the InvokeCommand for more detail about this type.

fn encode_unknown(&mut self, unknown: Vec<u8>)

Puts some unknown bytes into the buffer as it be. This expects that you satisfy following points:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

Parameters

unknown: Vec<u8>

Some unknown bytes to put into the buffer.

fn encode_chunk_data(&mut self, chunk_data: Option<ChunkData>)

Encodes some chunk data into the bytes. This expects that you satisfy following points:

If the extended timestamp is existing, you put this into just after the extended timestamp in the buffer.

Otherwise you put this into just after the chunk message header in the buffer.

If the chunk data is nothing, you put nothing into the buffer.

And you should use the method named "encode_*" corresponded to the type of the chunk pattern.
The correspondence of the chunk pattern to the method what should use when encode the chunk data is following:

ChunkData	Method

ChunkSize	encode_chunk_size()

BytesRead	encode_bytes_read()

PingData	encode_ping()

ServerBandwidth	encode_server_bandwidth()

ClientBandwidth	encode_client_bandwidth()

Audio	encode_audio()

Video	encode_video()

Notify	encode_notify()

Invoke	encode_invoke()

Otherwise	encode_unknown()

Parameters

chunk_data: Option<ChunkData>

The chunk data to put into the buffer. If chunk data to encode is nothing, pass the None. See the ChunkData for more detail about this type.

Loading content...