Module sheave::flv

The types for FLV file format.

We will be required to store sent Audio/Video data from the client because they must send the to other clients too. Therefore we must store several segments of Audio/Video data sent from the client just as much as needed. The FLV file format consists of following fields respectively:

The FLV header

The FLV bodies

The sequence of following fields:

The size of previous tag (unsigned 4 bytes number but the size of first tag must be 0)

The FLV tag

The FLV header

The FLV header consists of following fields respectively:

Signatures (array of unsigned 1 byte, total of 3 bytes)

Version (unsigned 1 byte number)

Reserved (unsigned 5 bits)

Audio flag (unsigned 1 bit)

Reserved (unsigned 1 bit)

Video flag (unsigned 1 bit)

Data offset (unsigned 4 bytes number, in byte)

Signatures

This is "FLV" as the string.

Version

The version for this file. For example, if this is 1, it indicates that this is FLV version 1.

Reserved of 5 bits

This shall be 0.

Audio flag

This indicates whether the audio data exists in this file.

Reserved of 1 bit

This shall be 0.

Video flag

This indicates whether the video data exists in this file.

Data offset

This is the offset to actual data. That is, the size of this header.

The FLV tag

The FLV tag consits of following fields respectively:

Reserved (unsigned 2 bits)

Filtered flag (unsigned 1 bit)

Tag type (unsigned 5 bits number)

Data size (unsigned 3 bytes number)

Timestamp (unsigned 3 bytes number)

Eextended timestamp (unsigned 1 byte number, in milliseconds)

This will be used when the timestamp has exceeded the limit for 3 bytes.

Note that this will be represented as upper 1 byte of the 4 bytes timestamp.

Stream id (unsigned 3 bytes)

Actual data

Audio

Video

ScriptData

Reserved

This shall be 0.

Filtered flag

This indicates whether the FLV bodies have been encrypted.

Tag type

Tha tag type consists of following values:

Number	Tag type

8	Audio

9	Video

18	ScriptData

Above numbers are identical with the message type id in the chunk message header. That is, the Audio tag indicates the Audio of chunk data, the Video tag indicates the Video of chunk data, and the ScriptData indicates the Notify of chunk data.

Data size

The actual data size. That is, this will be equal to what subturacts the total tag size to this header size.

Timestamp

The timestamp when this FLV tag was created. This field must count based on the timestamp of first FLV tag relatively, and the timestamp of first FLV tag must 0. This is counted in millisecond.

Extended timestamp

This will be used when the timestamp has exceeded 3 bytes limit, as the head 1 byte of 4 bytes timestamp.

Stream id

This must be input 0.

The encryption tag

The encryption tag consists of following fields respectively:

The additional header object

The encryption tag header

The filter params

The encryption body

The additional header object

If you use encryption, the additional header object will be sent as the Notify chunk with "|AdditionalHeader" message, as the first FLV element, and just after "onMetaData" of ScriptData tag. Note that the property is AMF's value. The additional header object consists of following property:

The encryption header object

The encryption header object

The encryption header object will be sent as the ScriptData tag named as "Encryption". And this consists of following properties:

Version (Number)

Method (String)

Some flags (Number)

Params (The standard encoding parameters object)

Sigformat/Signature (String/LongString)

Version

This indicate the Encryption header's version. The correspondence of the number to the version is following:

Number	Version

1	FMRMS v1.x products

2	Flash Access 2.0 products

Method

This shall be "Standard".

Some flags

This shall be 0.

Standard encoding parameters object

This consists of following properties:

Version (Number)

EncryptionAlgorithm (String)

EncryptionParams (The AES-CBC encryption parameters object)

KeyInfo (The key information object)

Version

This shall be 1.

EncryptionAlgorithm

This shall be "AES-CBC".

The AES-CBC encryption parameters object

This consists of following peoperty:

KeyLength (Number)

This shall be 16. (128 bits)

The key information object

This consists of following properties:

SubType (String)

If the encryption header's version is 1, this will be input "APS", otherwise this will be input "FlashAccessv2". APS is Adobe Policy Server. Online key agreement negotiation protocol. FlashAccessv2 is Flash Access 2.0 products. Online key retrieval protocol.

Data

If the encryption header's version is 1, this will be input the adobe policy server object, otherwise this will be input the flash access v2 object.

The flash access v2 object consists of a LongString value named as "MetaData". It's enceded by Base64, and will be used the DRM client to retrieve the decryption key.

The adobe policy server object couldn't be found the specification due to no longer produced by conforming applications.

Sigformat/Signature (String/LongString)

This hasn't written no document in the FLV file format specification.

The encryption tag header

The encryption tag header consists of following fields respectively:

Number of filters (unsigned 1 byte)

Filter name (String)

Length (unsigned 3 bytes)

Number of filters

This indicates the count of filters applied to the packet. This shall be 1.

Filter name (String)

This is a string for filter name. The name is following:

"Encryption" (if the version in the encryption header is 1)

"SE" (otherwise)

Length (unsigned 3 bytes)

This indicates the length of the filter params (in byte).

The filter params

The filter params consist of following fields respectively:

Encryption filter params

or

Selective encryption filter params

The encryption filter params

The encryption filter params consist of following field:

IV (array of unsigned 1 byte, total of 16 bytes)

IV

The initialization vector to be used for AES-CBC encryption.

The selective encryption filter params

The selective encryption filter params consist of following fields respectively:

Encrypted AU (unsigned 1 bit)

Reserved (unsigned 7 bits)

Encrypted AU

This is indicates whether the packet is encrypted.

Reserved

This shall be 0.

IV (array of unsigned 1 byte, total of 16 bytes)

The initialization vector to be used for AES-CBC encryption. This will be skipped unless the field of Encrypted AU is 1.

The encryption body

The encryption body consists of following fields respectively:

Content (array of unsigned 1 byte)

Padding (array of unsigned 1 byte)

Content

If the field of Encrypted AU is 0, this will be input the plaintext. If it is 1, this will be input the encrypted text. This length is equal to the input text length.

Padding (array of unsigned 1 byte)

This is the padding string for encrypting the content. This is also encrypted. This length is equal to actual padding string length. See RFC2630 for more detail about this.

The audio tag

The audio tag consists of following fields respectively:

The audio tag header.

The audio data.

The audio tag header.

The audio tag header consits of following fields respectively:

Sound format (unsigned 4 bits)

Sound rate (unsigned 2 bits)

Sound size (unsigned 1 bit)

Sound type (unsigned 1 bit)

AAC packet type (unsigned 1 byte)

Sound format

This indicates the audio file format. The correspondence of the number to the sound format is following:

Number	Sound format

0	Linear PCM (native endian)

1	ADPCM

2	MP3

3	Linear PCM (little endian)

4	Nellymoser 16kHz mono

5	Nellymoser 8kHz mono

6	Nellymoser

7	G.711 A-law logarithmic PCM

8	G.711 mu-law logarithmic PCM

9	Reserved

10	AAC

11	Speex

14	MP3 8kHz

15	Device-specific sound

Sound rate

This indicates the audio sampling rate. The correspondence of the number to the sound rate is following:

Number	Sound rate (in kHz)

0	5.5

1	11.0

2	22.0

3	44.0

Sound size

This indicates which the audio is either 8 bits or 16 bits. The correspondence of the number to the sound size is following:

Number	Sound size (in bit)

0	8

1	16

Sound type (unsigned 1 bit)

This indicates which the audio is either mono or stereo. The correspondence of the number to the sound type is following:

Number	Sound type

0	Mono

1	Stereo

AAC packet type (unsigned 1 byte)

This indicates which the audio is either AAC sequence header or AAC raw, if it is AAC. Therefore this will be skipped unless the codec is AAC. The correspondence of the number to the AAC packet type is following:

Number	AAC packet type

0	AAC sequence header

1	AAC raw

The audio data

This has depended on the audio codec for its byte format. See other documents about the audio codecs.

The video tag

The video tag consists of following fields respectively:

The video tag header

The video data

The video tag header

The video tag header consists of following fields respectively:

Frame type (unsigned 4 bits)

Codec id (unsigned 4 bits)

AVC packet type (unsigned 1 byte)

Composition time (unsigned 3 bytes)

Frame type

This indicates the video frame type. The correspondence of the number to the frame type is following:

Number	Frame type

1	key frame

2	inter frame

3	disposable inter frame

4	generated key frame

5	video info/command frame

Codec id (unsigned 4 bits)

This indicates the codec for video data. The correspondence of the number to the video codec is following:

Number	Video codec

2	Sorenson H.263

3	Screen video

4	On2 VP6

5	On2 VP6 with alpha channel

6	Screen video version 2

7	AVC

AVC packet type (unsigned 1 byte)

This indicate which the video is either AVC sequence header, AVC NALU or AVC end of sequence. Therefore this will be skipped unless the codec is AVC. The correspondence of the number to the AVC packet type is following:

Number	AVC packet type

0	AVC sequence header

1	AVC NALU

2	AVC end of sequence

Composition time (signed 3 bytes)

This indicates the time offset for the AVC codec. Therefore this will be skipped unless the codec is AVC, and will be input 0 unless the AVC packet type is 1 (NALU). See the ISO 14496-12 for more detail about the Composition time.

The video data

This has depended on the video codec for its byte format. See other documents about the video codecs.

The data tag

The data tag consists of what is identical to the MixedArray of AMF sent as the Notify chunk currently. That is, this consists of following fields respectively:

Name

ScriptData

Name

"onMetaData" (String)

ScriptData

Following name/value pairs. However these aren't input all necessarily.

Name	AMF data type

audiocodecid	Number

audiodatarate	Number

audiodelay	Number

audiosamplerate	Number

audiosamplesize	Number

canSeekToEnd	Boolean

creationdate	String

duration	Number

filesize	Number

framerate	Number

height	Number

stereo	Boolean

videocodecid	Number

videodatarate	Number

width	Number

See AmfData for more detail about the AMF.

Structs

AudioTag
The audio tag

AudioTagHeader
The audio tag header

DataTag
The Data tag

EncryptionTag
The encryption tag

EncryptionTagHeader
The encryption tag header

Flv
The FLV

FlvBody
The FLV file bodies

FlvHeader
The FLV header

FlvTag
The FLV tag

VideoTag
The video tag

VideoTagHeader
The video tag header

Enums

AacPacketType
The AAC packet type

AvcPacketType
The AVC packet type

Codec
The codec

FilterParams
The filter parameters

FlvData
The FLV data

FrameType
The frame type

SoundFormat
The sound format

SoundRate
The sound rate

SoundSize
The sound size

SoundType
The sound type

TagType
The tag type