Function sheave::run

pub fn run() -> IOResult<()>

Starts the server.

This takes following steps respectively:

Keeps the timestamp

Generates the TCP connection

Handles the RTMP chunks

Keeps the timestamp

First, Keeps the timestamp when started. All chunk will refer to this for counting the difference of the timestamp.

Generates the TCP connection

Generates the TCP connection as "127.0.0.1:1935" (currently). Then, waits for the client to connect to the server.

Handles the RTMP chunks

Handles the RTMP chunks. Note that in several chunks, sending plural responses will be required, or also sending no response will be required. See handlers.rs for more detail about this phase, and see messages.rs for more detail about each chunk.

Errors

When you got the ChunkLengthError:

The server couldn't read the chunk completely.

When you got the DigestVerificationError:

The HMAC-SHA256 digest didn't find in the C1 chunk.

When you got the SignatureDoesNotMatchError:

The HMAC-SHA256 signature in the C2 chunk didn't match with stored one in the server.

When you got the ChunkLengthError:

The format of some header or some chunk data is invalid.