Module sheave::handlers

The struct to handle RTMP

The RTMP will be required to exchange several chunks beforehand for streaming with audio/video data. Their exchanging chunks are following:

The RTMP handshake

The connect invocation

The releaseStream invocation

The FCPublish invocation

The createStream invocation

The publish invocation

Publishing audio/video data

The RTMP handshake

See the handshake.rs.

The connect invocation

We will exchange the information of the application each other. At this momennt, the server must send following responses respectively before sending its result:

The server-side bandwidth limit

The client-side bandwidth limit

The ping event (Stream Begin)

The chunk size limit

Note that somehow the server will be required to send a result of the connect invocation twice to FFmpeg. See the InvokeCommand and the NetConnectionCommand for more detail about the connect invocation. And see messages.rs for more detail about the server-side bandwidth limit, the client-side bandwidth limit, the ping event and the chunk size limit.

The releaseStream invocation

The client will send the identifier to indicate audio/video data, then the server will respond its result to the client. See the InvokeCommand and the NetConnectionCommand for more detail about the releaseStream invocation.

The FCPublish invocation

The client will send the same identifier as releaseStream, then the server will respond its result to the client. See the InvokeCommand and the FcPublishCommand for more detail about the FCPublish invocation.

The createStream invocation

The client will send a request to emit the message stream id, then the server will respond its result contained emitted message stream id to the client. See the InvokeCommand and the NetConnectionCommand for more detail about the createStream invocation.

The publish invocation

The client will send a message to tell starting to publish audio/video data, then the server will respond its result contained the server status to the client. At this moment, the server must send the ping event (Stream Begin) before sending its result. See the InvokeCommand and the NetStreamCommand for more detail about the publish invocation. And see messages.rs for more detail about the ping event.

Publishing audio/video data

The client will start to publish the audio/video data. At this moment, the server will receive a metadata of their audio/video data as the Notify chunk from the client at the first. See flv.rs for more detail about their auido/video data. And see the Metadata for more detail about the metadata.

Structs

LastChunk
The last sent/received chunk information

RtmpHandler
The RTMP handler

Enums

RtmpState
The patterns of RTMP