Module sheave::handshake

The structs/functions to do handshake for the RTMP layer.

The RTMP also needs to do handshake like the TCP.

The chunk format

The chunk what will send from server-side/client-side is following respectively.

S0/C0 (1 byte)

S1/C1 (1536 bytes)

S2/C2 (1536 bytes)

S0 and C0

Input the RTMP version.

3 (Default): The raw RTMP. This isn't encrypted neither the packet nor the network.

6, 8 and 9: The RTMPE (RTMP Encrypted). This encrypts just the packet. Its algorithm is following:

6: Uses just the Diffie-Hellman key exchange.

8: Uses the Diffie-Hellman key exchange with XTEA.

9: Uses the Diffie-Hellman key exchange with Blowfish.

However note that 6, 8 and 9 (that is, RTMPE) shouldn't use currently because these can be targetted from man-in-the-middle attack. RTMPE doesn't encrypt the network.

S1 and C1

This has four segments. Input following values in them respectively:

Timestamp (4 bytes)

Version (4 bytes)

Random Bytes (1536 bytes)

Timestamp

The timestamp what started the RTMP handshake. This hasn't been specified neither in second, in millisecond, nor in nanosecond. However this may be ok to input in-second timestamp due to just four bytes. This may set 0.

Version

In the official specification paper, this had been specified to input just four bytes zeroes. However this has currently been input the Adobe Media Server/the Flash Player version in many of the products already to being published. This is used to switch either to use HMAC-SHA256 or to use raw random bytes as them be, in the handshake.

The random bytes

The random bytes to use for the RTMP handshake. In the Adobe Media Server version and above 3/the Flash Player version and above 9, this has been each other imprinted HMAC-SHA256 digest. We will calculate the place where imprints its digest following formulas.

In the raw RTMP: (R[0] + R[1] + R[2] + R[3]) % 778 + 12

In the RTMPE: (R[764] + R[765] + R[766] + R[767]) % 728 + 776

Note where R is 1528 bytes randoms. That is, above formulas aren't considering the segments which the RTMP version and the Adobe Media Server/the Flash Player version.

The keys to use to get its digest are following.

Server: "Genuine Adobe Media Server 001"

Client: "Genuine Adobe Flash Player 001"

S2 and C2 (1536 bytes)

This is the chunk to echo each other the S1/C1 chunk. In the server, we will respond the C1 chunk. In the client, we will respond the S1 chunk. However in the Adobe Media Server version and above 3/the Flash Player version and above 9, we must imprint the HMAC-SHA256 signature at last 32 bytes. How to create its signature is following.

Find the digest imprinted in S1/C1. We can find it by the same way as what we calclated the place to imprint the digest in S1/C1.

Encrypt its digest by HMAC-SHA256.

Encrypt S1/C1 by HMAC-SHA256 except last 32 bytes.

Imprint its signature at last 32 bytes.

Note that the server will imprint the C1 by the server keys, and the client will imprint the S1 by the client keys.

The keys to use to get its signature is following.

Server: "Genuine Adobe Media Server 001" +
0xF0EEC24A8068BEE82E00D0D1029E7E576EEC5D2D29806FAB93B8E636CFEB31AE

Client: "Genuine Adobe Flash Player 001" +
0xF0EEC24A8068BEE82E00D0D1029E7E576EEC5D2D29806FAB93B8E636CFEB31AE

We will compare the signature which we got one and already imprinted one. If they haven't coincided, it will mean we've received some invalid handshake chunk. Note:

In the FFmpeg, above validation will fail because somehow it won't imprint its signature for the C2 chunk.

In the OBS (Open Broadcaster Software), above validation hasn't been needed in the first place because it has imprinted neither its digest nor its signature.

The Sending/Receiving sequence

We will send the S0/C0 chunk, the S1/C1 chunk, and the S2/C2 chunk following order.

The client will send the C0 chunk and the C1 chunk to the server.

When the server received the C0 chunk and the C1 chunk, the server will send the S0 chunk, the S1 chunk, and the S2 chunk to the client.

When the client received the S0 chunk, the S1 chunk and the S2 chunk, the client will validate its digest and its signature imprinted in the S2 chunk.
If they'are valid, the client will send the C2 chunk to the server.

When the server received the C2 chunk, the server will validate its digest and its signature imprinted in the C2 chunk.
If they're valid, the server will go to the phase of the application connection.

Structs

RtmpHandshake
The handshake validator for the RTMP