Module sheave::messages

The chunk message patterns

After doing the handshake, the server/the client will send actual messages each other. The message consists of following respectively:

The chunk basic header

The chunk message header

The extended timestamp

The chunk data

The chunk basic header

This will indicate the chunk stream id and the format of the chunk message header. This will hold following values respectively:

The format of the chunk message header (2 bits)

The chunk stream id (6 bits, 1 byte or 2 bytes)

The format of the chunk message header

The chunk message header will rely on this field for its length. The correspondence of the number to the chunk message header is following:

Number	Length (in bytes)

0	11

1	7

2	3

3	0

The chunk stream id

This will be used to store last chunks every this id. Note following points:

When this is below 64, input following the format bits as 6 bits.

When this is above 64, this will be input in next 1 byte.

Both (the server and the client) will regard this as what lessened just 64.

If this is 64 to 319, input 0 as 6 bits following the format bits.
This is Big Endian.

If this is 319 to 65599, input 1 as 6 bits following the format bits.
This is Little Endian.

The chunk message header

This will indicate the way to handle the chunk data. This will hold following values respectively:

Timestamp (3 bytes)

Message length (3 bytes)

Message type (1 byte)

Chunk message id (4 bytes)

Timestamp

The timestamp when this chunk is sent. Note following points:

If this will exceed 3 bytes, input this to the extended timestamp field instead.

In this case, input 0xFFFFFF (just the maximum value of 3 bytes) into this.

This can be 0 in some cases.

Message length

The length of the chunk data. Note following points:

This won't mean that is the total length of the chunk data.

This won't be considered to count the chunk header of format 3 contained in the chunk data.

We will be required to remove or to ignore it because it is contained at 1 byte after specified chunk size.

If we input total size contained the chunk header of format 3, many of products will output the error probably.

Message type

The type of the chunk data. The correspondence of the number to the chunk data is following:

Number	Message Type	Length (in byte)

1	Chunk size	4

2	Abort	4

3	Bytes read	4

4	Ping	Variable

5	Server bandwidth	4

6	Client bandwidth	5

8	Audio	Variable

9	Video	Variable

15	Notify (AMF3)	Variable

16	Shared object (AMF3)	Variable

17	Invoke (AMF3)	Variable

18	Notify (AMF0)	Variable

19	Shared object (AMF0)	Variable

20	Invoke (AMF0)	Variable

22	Metadata	Variable

Message stream id

The id to identify the user who sent this chunk. Note following points:

This can be 0 in some cases.

This must be emitted by the server when received Invoke(createStream).

This is Little Endian.

The patterns of the chunk message header every the format are following:

Field	Format 0	Format 1	Format 2	Format 3

Timestamp	✔	✔	✔	-

Message length	✔	✔	-	-

Message type	✔	✔	-	-

Message stream id	✔	-	-	-

The extended timestamp (4 bytes)

When the timestamp has exceeded 3 bytes, input it into this field instead. Note that mustn't input if it hasn't exceeded 3 bytes. If input in its case, the program can misunderstand as the chunk data.

The chunk data

The content of this chunk. The patterns are following:

Chunk size

This will indicate the size what will split the chunk data. The server/the client will be required to insert the chunk header of format 3 per who specified chunk size. In the official specification paper, this is specified that the most significant bit must be 0. (However we won't be required to care this normally because the message length will be represented at most 3 bytes.)

Abort

This will be input the chunk stream id which will abort to send/receive. When we received this chunk data, we are required to stop to send/receive the chunk data related to this chunk stream id. However in the FFmpeg and in the Open Broadcaster Software, both have never used this chunk data yet.

Bytes read

This will be input the total message length which has receive until now. The server/the client must send this chunk data each other whenever read byte size reached specified server bandwidth/client baandwidth.

Ping

This will be used to confirm current state for the server/the client each other. The values to be required to input are following respectively:

Event type (2 bytes)

This will indicate the kinds of data following this bytes. The correspondence of the number to actual event data is following:

Number	Event type	Length (in byte)

0	Stream begin	4

1	Stream EOF	4

2	Stream dry	4

3	Buffer length	8

4	Stream is recorded	4

6	Ping	4

7	Pong	4

26	SWF verification request	0

27	SWF verification response	42

31	Buffer empty	4

32	Buffer ready	4

Event data

This will be input actual ping data.

Stream begin, Stream EOF, Stream dry, Stream is recorded, Buffer empty and Buffer ready.

This will be input only the message stream id (4 bytes). In the phase of the application connection, the chunk message id will be 0 inevitably.

Buffer length

This will be input the message stream id (4 bytes) and the buffer length (4 bytes). Note that this has differed from the chunk size. This has indicated the size to send/to receive the data in millisecond.

Ping and Pong

This will be input the timestamp when the server emitted this event.

SWF verification request

This will be input no data.

SWF verification response

This will be input the HMAC-SHA256 signature what generated from the SWF data. Note that we will be required to input following bytes respectively before its signature:

1 (1 byte)

1 (1 byte)

the size of raw SWF data. (4 bytes)

the size of raw SWF data. (4 bytes)

Server bandwidth

This will be input the limit of the server side bandwidth.

Client bandwidth

This will be input the limit of the client side bandwidth and the limit type. The limit type is following:

Hard

Soft

Dynamic

Hard

This indicates that the receiver should limit the bandwidth to this.

Soft

This indicates that the receiver should limit the bandwidth to either this value or already received one, whichever is smaller.

Dynamic

This indicates that the receiver should regard this as the limit type of hard if the previous limit type was hard, otherwise ignore this message.

The correspondence of the number to the limit type is following:

Number	Limit type

0	Hard

1	Soft

2	Dynamic

Audio

This will be input the audio data.

Video

This will be input the video data.

Note: the format for audio/video has differed every codec.

Notify

Currently, this will be input the metadata of audio/video data.

Shared Object

This will be input the information to share with the client or other server instances. However the client side applications published as the OSS haven't implemented this chunk data yet.

Invoke

This will be input the information to need what the server/the client application each will succeed to connect.

Metadata

This will be input the data to aggregate the audio, the video, or the notify(metadata) chunks.

Structs

Argument
This is the optional argumeent in the connect of the NetConnection command. This hasn't been checked its exsitence yet.

BasicHeader
The chunk basic header

ByteBuffer
The byte buffer to encode/decode the Rust's data/the bytes

Chunk
The chunk

CommandObject
The command object

InfoObject
The information object

MetaData
The metadata of FLV

Enums

AmfData
The AMF data

AmfDataType
The AMF types

AudioCodec
The audio codec patterns

ChunkData
The chunk data

ChunkId
The chunk stream id

ConnectStatus
The status of the connect of the NetConnection command

FcPublishCommand
The FCPublish command

InfoLevel
The information lebel for InfoObject

InvokeCommand
The command of the invoke chunk

LimitType
The client bandwidth limit type

MessageFormat
The format of the chunk message header

MessageHeader
The chunk message header

MessageType
The message types of the chunk message header

NetConnectionCommand
The NetConnection command.

NetConnectionResult
The result of the NetConnection command.

NetConnectionStatus
The status of the NetConnection command

NetStreamCommand
The NetStream command

NetStreamStatus
The status of the NetStream command

NotifyCommand
The notify command

ObjectEncoding
The object encoding pattern

PingData
The ping event data

PingType
The ping event type

PlayType
The publish pattern

PublishStatus
The status of the publish of the NetStream command

Status
The status code

VideoCodec
The video codec patterns

VideoFunction
The video function pattern

Functions

convert_audio_codecs_into_flags
Converts the audio codecs into the 16 bits flags.

convert_video_codecs_into_flags
Converts the video codecs into the 8 bits flags.

detect_audio_codecs
Converts the 16 bits flags for the audio codecs into the Vec<AudioCodec>.

detect_video_codecs
Converts the 8 bits flags for the video codecs into the Vec<VideoCodec>.