Enum sheave::messages::AudioCodec

#[repr(u16)]
pub enum AudioCodec {
    None,
    Adpcm,
    Mp3,
    Intel,
    Unused,
    Nerry8,
    Nerry,
    G711a,
    G711u,
    Nerry16,
    Aac,
    Speex,
    All,
}

The audio codec patterns

This indicates the audio codec to input in the command object. The correspondence of the number to this enum is following:

Number	AudioCodec

0x0001	None

0x0002	Adpcm

0x0004	Mp3

0x0008	Intel

0x0010	Unused

0x0020	Nerry8

0x0040	Nerry

0x0080	G711a

0x0100	G711u

0x0200	Nerry16

0x0400	Aac

0x0800	Speex

0x0fff	All

This enum and the u16 value can convert into each other because this has implemented the From<u16> and has set the #[repr(u16)] attribute.

Variants

None

Adpcm

Mp3

Intel

Unused

Nerry8

Nerry

G711a

G711u

Nerry16

Aac

Speex

All