Enum sheave::messages::NetConnectionCommand

pub enum NetConnectionCommand {
    Connect {
        argument: Option<Argument>,
        transaction_id: u64,
        command_object: CommandObject,
    },
    ConnectResult {
        result: NetConnectionResult,
        transaction_id: u64,
        properties: HashMap<String, AmfData>,
        information: InfoObject,
    },
    ReleaseStream {
        transaction_id: u64,
        play_path: String,
    },
    ReleaseStreamResult {
        result: NetConnectionResult,
        transaction_id: u64,
    },
    CreateStream {
        transaction_id: u64,
    },
    CreateStreamResult {
        result: NetConnectionResult,
        message_id: u32,
        transaction_id: u64,
    },
}

The NetConnection command.

This consists of following patterns:

Connect

ConnectResult

ReleaseStream

ReleaseStreamResult

CreateStream

CreateStreamResult

Connect

This consists of following data:

Field	Type

transactionn_id	u64

command_object	CommandObject

argument	Option<Argument>

transaction_id

The number to indicate the step of current invoking. This is converted from the AMF's Number types.

command_object

See the CommandObject.

argument

Currently, this will be the None inevitably. This has been defined in the specification but hasn't been able to check its existence yet. Because of the above causes, the Argument type has defined like the unit and has been contained in the Option type currently.

ConnectResult

This consists of following data:

Field	Type

result	NetConnectionResult

transaction_id	u64

properties	HashMap<String, AmfData>

information	InfoObject

result

See the NetConnectionResult.

transaction_id

This is the same value as the Connect's one.

properties

The HashMap of AmfData associated with the property name.

information

See the InfoObject.

ReleaseStream

This consists of following data:

Field	Type

transaction_id	u64

play_path	String

transaction_id

This is the same format as the Connect's one.

play_path

A part of the URL to be requested from the client when the TCP connection. i.e. playpath in rtmp://example.com/appName/playpath

This can start with the prefix of "mp4:", etc.

ReleaseStreamResult

This consists of following data:

Field	Type

result	NetConnectionResult

transaction_id	u64

result

See the NetConnectionResult.

transaction_id

This is the same value as the ReleaseStream's one.

CreateStream

This consists of following data:

Field	Type

transaction_id	u64

transaction_id

This is the same format as the ReleaseStream's one.

CreateStreamResult

This consists of following data:

Field	Type

result	NetConnectionResult

transaction_id	u64

message_id	u32

result

See the NetConnectionResult.

transaction_id

The same value as the CreateStream's one.

message_id

The number to identify its user. This will be used in following parts:

The chunk message header's message stream id

The ping's event data

Or some implementations to identify the user

Variants

Connect

Fields of Connect

argument: Option<Argument>transaction_id: u64command_object: CommandObject

ConnectResult

Fields of ConnectResult

result: NetConnectionResulttransaction_id: u64properties: HashMap<String, AmfData>information: InfoObject

ReleaseStream

Fields of ReleaseStream

transaction_id: u64play_path: String

ReleaseStreamResult

Fields of ReleaseStreamResult

result: NetConnectionResulttransaction_id: u64

CreateStream

Fields of CreateStream

transaction_id: u64

CreateStreamResult

Fields of CreateStreamResult

result: NetConnectionResultmessage_id: u32transaction_id: u64