Struct sheave::messages::ByteBuffer

pub struct ByteBuffer { /* fields omitted */ }

The byte buffer to encode/decode the Rust's data/the bytes

This consists of following data:

Field	Type

offset	usize

len	usize

bytes	Vec<u8>

This is implemented to premise that will dispose after receiving/sending the chunk. Therefore this is considered just to append the bytes, not to overwrite its bytes.

Methods

impl ByteBuffer

pub fn new(bytes: Vec<u8>) -> Self

Constructs a new ByteBuffer.

Parameters

bytes: Vec<u8>

The bytes to decode/encode. When we start to encode, pass to an empty Vec.

pub fn clear(&mut self)

Clear this byte buffer. This sets 0 to the offset, sets 0 to the len, and constructs a new Vec to the bytes again.

pub fn offset(&self) -> usize

Returns the offset.

pub fn offset_to(&mut self, offset: usize)

Forwards the offset.

Parameters

offset: usize

The offset to forward.

Note that we should increase the offset when have decoded some byte.

pub fn len(&self) -> usize

Returns the length.

pub fn add_len(&mut self, len: usize)

Adds the length.

Parameters

len: usize

The length to add.

Note that we should increase the length when have encoded some byte.

pub fn bytes(&self) -> &Vec<u8>

Returns all byte.

pub fn bytes_mut(&mut self) -> &mut Vec<u8>

Returns all byte as mutable.