Struct sheave::messages::CommandObject

pub struct CommandObject { /* fields omitted */ }

The command object

This consists of following data:

Field	Type

fpad	Option<bool>

object_encoding	Option<ObjectEncoding>

video_function	Option<VideoFunction>

video_codec	Option<Vec<VideoCodec>>

audio_codec	Option<Vec<AudioCodec>>

app	Option<String>

command_type	Option<String>

flash_ver	Option<String>

swf_url	Option<String>

tc_url	Option<String>

page_url	Option<String>

These fields won't be contained all necessarily. Actually, the command object to be sent from the FFmpeg will be contained just following fields:

app

command_type

flash_ver

tc_url

Because of actual command object is so, all field has been contained in the Option type. This type and the HashMap<String, AmfData> type can convert into each other because this has implemented the From<HashMap<String, AmfData>> and the From<CommandObject>.

Methods

impl CommandObject

pub fn new() -> Self

Constructs a new CommandObject.

pub fn set_fpad(&mut self, fpad: Option<bool>)

Sets the fpad.

Parameters

fpad: Option<bool>

A bool value converted from AMF's Boolean. If you set no value, pass the None.

pub fn set_object_encoding(&mut self, object_encoding: Option<f64>)

Sets the object encoding.

Parameters

object_encoding: Option<f64>

A f64 value converted from AMF's Number. This indicates the AMF version which its user uses. If you set no value, pass the None. See the ObjectEncoding for more detail about the object encoding.

pub fn set_video_function(&mut self, video_function: Option<f64>)

Sets the video function.

Parameters

video_function: Option<f64>

A f64 value converted from AMF's Number. This indicates the special video function in FLV. If you set no value, pass the None. See the VideoFunction for more detail about the video function.

pub fn set_video_codecs(&mut self, video_codecs: Option<f64>)

Sets the video codecs.

Parameters

video_codecs: Option<f64>

A f64 value converted from AMF's Number. This indicates supported video codecs. This can specify plural codecs because will be used as a 64 bits flag. If you set no value, pass the None. See the VideoCodec for more detail about the video codecs flag.

pub fn set_audio_codecs(&mut self, audio_codecs: Option<f64>)

Sets the audio codecs.

Parameters

audio_codecs: Option<f64>

A f64 value converted from AMF's Number. This indicates supported audio codecs. This can specify plural codecs because will be used as a 64 bits flag. If you set no value, pass the None. See the AudioCodec for more detail about the audio codecs flag.

pub fn set_app(&mut self, app: Option<String>)

Sets the application name.

Parameters

app: Option<String>

A String value converted from AMF's String (AmfData::String). This can be "". (that is, an empty string) If you set no value, pass the None.

pub fn set_command_type(&mut self, command_type: Option<String>)

Sets the command type.

Parameters

command_type: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.

pub fn set_flash_ver(&mut self, flash_ver: Option<String>)

Sets the Flash Player version.

Parameters

flash_ver: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.

pub fn set_swf_url(&mut self, swf_url: Option<String>)

Sets the URL of the video player application connecting with the client.

Parameters

swf_url: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.

pub fn set_tc_url(&mut self, tc_url: Option<String>)

Sets the URL of the server.

Parameters

tc_url: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.

pub fn set_page_url(&mut self, page_url: Option<String>)

Sets the URL of the web page where the video player application was loaded.

Parameters

page_url: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.