Struct sheave::messages::InfoObject

pub struct InfoObject { /* fields omitted */ }

The information object

This consists of following data currently:

Field	Type

object_encoding	Option<ObjectEncoding>

code	Option<Status>

level	Option<String>

details	Option<String>

description	Option<String>

This won't also be contained all field necessarily as the CommandObject type is so.

In common:

code

level

description

In just the result of the connect of the NetConnection command:

object_encoding

In just the onStatus of the publish of the NetStream command:

details

This type and the HashMap<String, AmfData> type can convert into each other because this has implemented the From<HashMap<String, AmfData>> and the From<InfoObject>

Methods

impl InfoObject

pub fn new() -> Self

Constructs a new InfoObject.

pub fn set_object_encoding(&mut self, object_encoding: Option<f64>)

Sets the object encoding.

Parameters

object_encoding: Option<f64>

A f64 value converted from AMF's Number. This indicates the AMF version which its user uses. If you set no value, pass the None. See the ObjectEncoding for more detail about the object encoding.

pub fn set_level(&mut self, level: Option<String>)

Sets the level.

Parameters

level: Option<String>

A String value converted from AMF's String (AmfData::String). This indicates the information level which the server will respond. You will be required to pass one of following values:

status

warning

error

If you set no value, pass the None. See the InfoLevel for more detail about the level.

pub fn set_code(&mut self, code: Option<String>)

Sets the code.

Parameters

code: Option<String>

A String value converted from AMF's String (AmfData::String). This indicates the status code which the server will respond. You will be required to pass one of following values:

NetConnection

Connect

Success

NetStream

Publish

Start

i.e. Either "NetConnection.Connect.Success" or "NetStream.Publish.Start". See the Status for more detail about the code.

pub fn set_details(&mut self, details: Option<String>)

Sets the details.

Parameters

details: Option<String>

A String value converted from AMF's String (AmfData::String). If you set no value, pass the None.

pub fn set_description(&mut self, description: Option<String>)

Sets the description.

Parameters

description: Option<String>

A String value converted from AMF's string (AmfData::String). If you set no value, pass the None.